+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## 0.0.10
+
+Change call widget.
+
+Before -> call via method
+example: DButton.circle()
+
+After -> call via class
+example: DButtonCircle
+
+## 0.0.3
+
+- change format doc comment on button shadow
+
+## 0.0.2
+
+- Add documentation
+- Add Button Border
+
 ## 0.0.1
 
 DButton package :
@@ -10,8 +29,3 @@ Child can be filled with any widget as long as there is sufficient space.
 But, for DButton.circle() just to use for layout circle.
 
 You can modify all the buttons above to your liking and be as creative as possible
-
-## 0.0.2
-
-- Add documentation
-- Add Button Border